requirements.txt
mcp
google-genai
python-dotenv
requests